# swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "The Microservices API of the USupport Platform"
  },
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "AnyAdminBearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter into the field the word 'Bearer' following by space and JWT"
    },
    "CountryAdminBearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter into the field the word 'Bearer' following by space and JWT"
    },
    "GlobalAdminBearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter into the field the word 'Bearer' following by space and JWT"
    },
    "AnyUserBearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter into the field the word 'Bearer' following by space and JWT"
    },
    "ClientBearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter into the field the word 'Bearer' following by space and JWT"
    },
    "ProviderBearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter into the field the word 'Bearer' following by space and JWT"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get Current User",
        "parameters": [
          {
            "name": "x-language-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the language"
          },
          {
            "name": "x-country-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the country"
          }
        ],
        "responses": {
          "200": {
            "description": "User Data Object"
          },
          "401": {
            "description": "User Not Authorised"
          },
          "404": {
            "description": "User Not Found"
          }
        },
        "security": [
          {
            "AnyUserBearer": []
          }
        ]
      }
    },
...
    "/user/countries/by-alpha-2-code": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get a specific country by its alpha-2 code",
        "parameters": [
          {
            "name": "x-language-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the language"
          },
          {
            "name": "x-country-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the country"
          }
        ],
        "responses": {
          "200": {
            "description": "Country Data Object"
          },
          "404": {
            "description": "Country Not Found"
          }
        }
      }
    },
    "/user/work-with": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get all work with areas",
        "parameters": [
          {
            "name": "x-country-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the country"
          }
        ],
        "responses": {
          "200": {
            "description": "Work With Data Object"
          }
        }
      }
    },
    "/user/provider/signup": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Create provider user account",
        "parameters": [
          {
            "name": "x-language-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the language"
          },
          {
            "name": "x-country-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the country"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "John"
                },
                "patronym": {
                  "type": "string",
                  "example": "Johny"
                },
                "surname": {
                  "type": "string",
                  "example": "Doe"
                },
                "nickname": {
                  "type": "string",
                  "example": "JD123"
                },
                "email": {
                  "type": "string",
                  "example": "john.doe@email.com"
                },
                "phonePrefix": {
                  "type": "string",
                  "example": "+44"
                },
                "phone": {
                  "type": "string",
                  "example": "1234567890"
                },
                "specializations": {
                  "type": "array",
                  "example": [
                    "psychologist",
                    "coach"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "street": {
                  "type": "string",
                  "example": "Some Street"
                },
                "city": {
                  "type": "string",
                  "example": "Another City"
                },
                "postcode": {
                  "type": "string",
                  "example": "1234"
                },
                "education": {
                  "type": "array",
                  "example": [
                    "Education 1",
                    "Education 2"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "sex": {
                  "type": "string",
                  "example": "unspecified"
                },
                "consultationPrice": {
                  "type": "number",
                  "example": 60
                },
                "description": {
                  "type": "string",
                  "example": "Some Long Description Here..."
                },
                "workWithIds": {
                  "type": "array",
                  "example": [
                    "22e3b2f6-5c95-4044-b444-592b5d41338a",
                    "ccd6a85d-ab7d-4700-953d-cda0775f37e5"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "languageIds": {
                  "type": "array",
                  "example": [
                    "69f03082-ee81-4a11-a7a2-84e82bd54369",
                    "2dc1092c-a13d-4d55-9b1f-81d3b3e974c1"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "name",
                "surname",
                "email"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "New Provider Data Object"
          },
          "401": {
            "description": "No Permissions"
          },
          "409": {
            "description": "Email Already Used"
          }
        }
      }
    },
    "/user/signup": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Create new client user account",
        "parameters": [
          {
            "name": "x-language-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the language"
          },
          {
            "name": "x-country-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the country"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "John"
                },
                "surname": {
                  "type": "string",
                  "example": "Doe"
                },
                "nickname": {
                  "type": "string",
                  "example": "JD123"
                },
                "email": {
                  "type": "string",
                  "example": "john.doe@email.com"
                },
                "userAccessToken": {
                  "type": "string",
                  "example": "b1igkLasNQ"
                },
                "sex": {
                  "type": "string",
                  "example": "unspecified"
                },
                "yearOfBirth": {
                  "type": "number",
                  "example": 2000
                },
                "urbanRural": {
                  "type": "string",
                  "example": "urban"
                }
              },
              "required": [
                "nickname"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "New Client Data Object"
          },
          "409": {
            "description": "Email Already Used"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Login user with email or user access token",
        "parameters": [
          {
            "name": "x-language-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the language"
          },
          {
            "name": "x-country-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the country"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userType": {
                  "type": "string",
                  "example": "client"
                },
                "email": {
                  "type": "string",
                  "example": "john.doe@email.com"
                },
                "userAccessToken": {
                  "type": "string",
                  "example": "b1igkLasNQ"
                },
                "password": {
                  "type": "string",
                  "example": "SomePass123"
                }
              },
              "required": [
                "userType",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client/Provider Access and Refresh Tokens"
          },
          "404": {
            "description": "Incorrect Password"
          }
        }
      }
    },
...
    "/user/user-access-token": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Generate new user access token for signup",
        "parameters": [
          {
            "name": "x-language-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the language"
          },
          {
            "name": "x-country-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the country"
          }
        ],
        "responses": {
          "200": {
            "description": "User Access Token Object"
          }
        }
      }
    },
    "/user/refresh-token": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Refresh JWT access token",
        "parameters": [
          {
            "name": "x-language-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the language"
          },
          {
            "name": "x-country-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the country"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "type": "string",
                  "example": "22e3b2f6-5c95-4044-b444-592b5d41338a"
                }
              },
              "required": [
                "refreshToken"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "New Access and Refresh Tokens"
          },
          "401": {
            "description": "Refresh Token Not Valid"
          }
        }
      }
    },
    "/user/password": {
      "patch": {
        "tags": [
          "User"
        ],
        "description": "Change user",
        "parameters": [
          {
            "name": "x-language-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the language"
          },
          {
            "name": "x-country-alpha-2",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Alpha 2 code of the country"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "oldPassword": {
                  "type": "string",
                  "example": "OldPass123"
                },
                "newPassword": {
                  "type": "string",
                  "example": "NewPass123"
                }
              },
              "required": [
                "oldPassword",
                "newPassword"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success Status"
          },
          "401": {
            "description": "User Not Authorised"
          },
          "404": {
            "description": "Incorrect Password"
          }
        },
        "security": [
          {
            "AnyUserBearer": []
          }
        ]
      }
    },
...
  }
}